Do not resolve symlinks in filepath
If editing a file by opening its symlink, the output file created by
pandoc uses the symlink name rather than actual file name. If the
resolve() function is used, the plugin looks for output with the file
basename instead of the symlink basename, so do not resolve the edited
file's name.